Create Gbuscount
There exist some cities that are built along a straight road. The cities are numbered 1,2,3,… from left to right.

There are N GBuses that operate along this road. For each GBus, the range of cities that it serves is provided: the i-th gBus serves the cities with numbers between Ai and Bi, inclusive.

We are interested in a particular subset of P cities. For each of those cities, we need to find out how many GBuses serve that particular city.